requirements.txt
# GUI
PyQt6>=6.4.0

# 数据库
SQLAlchemy>=2.0.0

# AI服务
requests>=2.28.0
aiohttp>=3.8.0

# 文档导出
python-docx>=0.8.11
reportlab>=3.6.12

# 工具
python-dotenv>=0.19.0
pyyaml>=6.0.0
loguru>=0.6.0

# 开发工具
black>=22.3.0
isort>=5.10.1
mypy>=0.982
pytest>=7.1.3
pytest-asyncio>=0.19.0
pytest-qt>=4.2.0
